#GAM-23030 #GAM-23022 #GAM-22977 #GAM-22978 #GAM-22983 Vesting & NFTStaking & Verichains fix

+++ contracts/JamDistribute.sol
@@ -4,11 +4,10 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/utils/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/security/ReentrancyGuard.sol";
-import "./HasNoEther.sol";
-import "./DateTime.sol";
+import "./utils/HasNoEther.sol";
+import "./utils/DateTime.sol";
 
 contract JamDistribute is ReentrancyGuard, HasNoEther {
-
     using SafeMath for uint256;
     using SafeERC20 for IERC20;
 
@@ -21,49 +20,63 @@ contract JamDistribute is ReentrancyGuard, HasNoEther {
     // Store rewards of user link to specific type of token
     mapping(address => mapping(address => uint256)) private rewards;
 
-    // Total Amount of reward of a specific token, 
+    // Total Amount of reward of a specific token,
     // just for the sake of no looping
     mapping(address => uint256) private totalRewards;
 
     // Mapping store state of supported distribute token in the contract
     mapping(address => bool) private supportedToken;
-    
+
     /* ========== CONSTRUCTOR ========== */
-    constructor(
-        address _newOwner
-    ) {
+    constructor(address _newOwner) {
         transferOwnership(_newOwner);
     }
 
-
     /* ======== Native Function ========= */
-    function viewReward(address tokenAddr, address addr) public view returns (uint256){
+    function viewReward(address tokenAddr, address addr)
+        public
+        view
+        returns (uint256)
+    {
         return rewards[tokenAddr][addr];
     }
 
-    function addRewards(address tokenAddr, address[] memory addrs, uint256[] memory amount) external
+    function addRewards(
+        address tokenAddr,
+        address[] memory addrs,
+        uint256[] memory amount
+    )
+        external
         nonReentrant
         onlyDistributor(tokenAddr)
         enoughBalance(tokenAddr, amount)
         onlySupportedToken(tokenAddr)
     {
-        require(addrs.length == amount.length, "addrs and amount does not same length");
+        require(
+            addrs.length == amount.length,
+            "addrs and amount does not same length"
+        );
         uint256 addedRewardAmount = _calSumAmount(amount);
         uint256 today = DateTime.toDateUnit(block.timestamp);
-        for (uint i=0; i < addrs.length; i++) {
-            rewards[tokenAddr][addrs[i]] = (rewards[tokenAddr][addrs[i]]).add(amount[i]);
+        for (uint256 i = 0; i < addrs.length; i++) {
+            rewards[tokenAddr][addrs[i]] = (rewards[tokenAddr][addrs[i]]).add(
+                amount[i]
+            );
             if (latestAddRewardDate[tokenAddr][addrs[i]] >= today) {
                 revert("reward already added for today");
             }
             latestAddRewardDate[tokenAddr][addrs[i]] = today;
         }
-        totalRewards[tokenAddr] = totalRewards[tokenAddr].add(addedRewardAmount);
+        totalRewards[tokenAddr] = totalRewards[tokenAddr].add(
+            addedRewardAmount
+        );
     }
 
-    function updateDistributors(address tokenAddr, address distributor, bool ok) external 
-        onlyOwner
-        onlySupportedToken(tokenAddr) 
-    {
+    function updateDistributors(
+        address tokenAddr,
+        address distributor,
+        bool ok
+    ) external onlyOwner onlySupportedToken(tokenAddr) {
         allowedDistributors[tokenAddr][distributor] = ok;
         if (ok) {
             emit AddDistributor(tokenAddr, distributor);
@@ -72,7 +85,8 @@ contract JamDistribute is ReentrancyGuard, HasNoEther {
         }
     }
 
-    function updateSupportedToken(address tokenAddr, bool ok) external
+    function updateSupportedToken(address tokenAddr, bool ok)
+        external
         onlyOwner
     {
         supportedToken[tokenAddr] = ok;
@@ -83,7 +97,8 @@ contract JamDistribute is ReentrancyGuard, HasNoEther {
         }
     }
 
-    function getReward(address tokenAddr) external
+    function getReward(address tokenAddr)
+        external
         haveReward(tokenAddr)
         onlySupportedToken(tokenAddr)
     {
@@ -96,30 +111,39 @@ contract JamDistribute is ReentrancyGuard, HasNoEther {
         } else {
             IERC20(tokenAddr).safeTransfer(msg.sender, rewardAmount);
         }
-        rewards[tokenAddr][msg.sender] = rewards[tokenAddr][msg.sender].sub(rewardAmount);
+        rewards[tokenAddr][msg.sender] = rewards[tokenAddr][msg.sender].sub(
+            rewardAmount
+        );
         totalRewards[tokenAddr] = totalRewards[tokenAddr].sub(rewardAmount);
     }
 
-    function _calSumAmount(uint256[] memory amount) private pure returns (uint256) {
+    function _calSumAmount(uint256[] memory amount)
+        private
+        pure
+        returns (uint256)
+    {
         uint256 total = 0;
-        for(uint i = 0; i < amount.length; i++) {
+        for (uint256 i = 0; i < amount.length; i++) {
             total = total.add(amount[i]);
         }
         return total;
     }
- 
+
     fallback() external payable {}
 
     receive() external payable {}
-    
+
     /* ======== Modfier ========= */
 
     modifier onlyDistributor(address _token) {
-        require(allowedDistributors[_token][msg.sender], "only distributor can take this action");
+        require(
+            allowedDistributors[_token][msg.sender],
+            "only distributor can take this action"
+        );
         _;
     }
 
-    modifier onlySupportedToken(address _token){
+    modifier onlySupportedToken(address _token) {
         require(supportedToken[_token], "unsupported token");
         _;
     }
@@ -132,18 +156,30 @@ contract JamDistribute is ReentrancyGuard, HasNoEther {
         } else {
             thisAccountBalance = IERC20(tokenAddr).balanceOf(address(this));
         }
-        require(totalRewards[tokenAddr].add(total) <= thisAccountBalance, "balance not enough");
+        require(
+            totalRewards[tokenAddr].add(total) <= thisAccountBalance,
+            "balance not enough"
+        );
         _;
     }
 
     modifier haveReward(address tokenAddr) {
-        require(rewards[tokenAddr][msg.sender] > 0, "you don't have any reward");
+        require(
+            rewards[tokenAddr][msg.sender] > 0,
+            "you don't have any reward"
+        );
         _;
     }
 
     /* ======== Event ========= */
-    event AddDistributor(address indexed tokenAddr, address indexed distributor);
-    event RemoveDistributor(address indexed tokenAddr, address indexed distributor);
+    event AddDistributor(
+        address indexed tokenAddr,
+        address indexed distributor
+    );
+    event RemoveDistributor(
+        address indexed tokenAddr,
+        address indexed distributor
+    );
     event AddToken(address indexed tokenAddr);
     event RemoveToken(address indexed tokenAddr);
 }
